% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{helper}
\alias{helper}
\alias{bpwhisker.l}
\alias{bpwhisker.u}
\alias{ci.header}
\alias{mod.ref}
\title{Helper Functions}
\usage{
bpwhisker.l(x)

bpwhisker.u(x)

ci.header(digits)

mod.ref(tp, ref.dp, digits = 4, model, time.unit)
}
\arguments{
\item{x}{Numeric vector}

\item{digits}{An integer indicating the decimal points for the output.}

\item{tp, ref.dp}{Numeric vector of time points \code{tp} and their
corresponding mean dissolution profiles \code{ref.dp}.}

\item{model}{Strings of model names. Currently only 'Weibull' and
'first-order' models are supported.}

\item{time.unit}{Character strings indicating the unit of time. It should
be either \code{"min"} for minute or \code{"h"} for hour. It is mainly
used for checking CV rules and making plot. @seealso \code{\link[=calcf2]{calcf2()}}.}
}
\description{
Helper Functions
}
