% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_shah1998.R
\docType{data}
\name{shah1998}
\alias{shah1998}
\title{Dissolution data from the article of Shah et al 1998}
\format{
A list of 6 data frames. Each data frame is a set of dissolution
profiles with the format: the first column is time, the rest columns
are individual profiles.
\describe{
\item{ref}{dissolution of reference batch}
\item{test1}{dissolution of test batch 1}
\item{test2}{dissolution of test batch 2}
\item{test3}{dissolution of test batch 3}
\item{test4}{dissolution of test batch 4}
\item{test5}{dissolution of test batch 5}
}
}
\source{
Shah VP, Tsong Y, Sathe P, Liu JP. In vitro dissolution profile
comparison--statistics and analysis of the similarity factor, f2.
Pharm Res. 1998 Jun;15(6):889-96. doi: 10.1023/a:1011976615750.
}
\usage{
shah1998
}
\description{
Dissolution data of reference and 5 batches of test products published
in the article of Shah et al, Pharm. Res.1998, 15(6), 889--896.
}
\keyword{datasets}
