% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootf2.R
\name{bootf2}
\alias{bootf2}
\title{Estimate 90\% Confidence Intervals of f2 Using Bootstrap Methodology}
\usage{
bootf2(test, ref, path.in, file.in, path.out, file.out,
       n.boots = 10000L, seed = 306, digits = 2L, alpha = 0.05,
       regulation = c("EMA", "FDA", "WHO"), min.points = 1L,
       both.TR.85 = FALSE, print.report = TRUE,
       report.style = c("concise", "intermediate", "detailed"),
       f2.type = c("all", "est.f2", "exp.f2", "bc.f2",
                   "vc.exp.f2", "vc.bc.f2"),
       ci.type = c("all", "normal", "basic", "percentile",
                   "bca.jackknife", "bca.boot"),
       quantile.type = c("all", 1:9, "boot"),
       jackknife.type = c("nt+nr", "nt*nr", "nt=nr"),
       time.unit = c("min", "h"), output.to.screen = TRUE,
       sim.data.out = FALSE)
}
\arguments{
\item{test, ref}{Data frames of dissolution profiles of test and reference
product if \code{path.in} and \code{file.in} are not specified; otherwise,
they should be character strings indicating the worksheet names of the
Excel file where the dissolution data is saved. Required format:
the first column should be time and the rest columns are dissolution
data of each unit. See Input/Output in Details.}

\item{path.in, file.in, path.out, file.out}{Character strings of input and
output directories and file names. See Input/Output in Details.}

\item{n.boots}{An integer indicating the number of bootstrap samples.}

\item{seed}{Numeric seed value for reproducibility. If missing, a random
seed will be generated for reproducibility purpose.}

\item{digits}{An integer indicating the decimal points for the output.}

\item{alpha}{A numeric value between 0 and 1.
\eqn{(1-2\times \alpha)\times 100}{(1 - 2*alpha)*100}
confidence interval will be estimated.}

\item{regulation}{Character strings indicating regulatory guidelines.
See Regulation in Details.}

\item{min.points}{An integer indicating the minimum time points to be used
to calculate f2. See Regulation in Details.}

\item{both.TR.85}{Logical. If \code{TRUE}, the old (incorrect)
interpretation (that both the test and reference should release more than
85\%) will be used for f2 calculation when \code{regulation = 'FDA'}.
See Regulation in Details.}

\item{print.report}{Logical. If \code{TRUE}, a plain text report will be
produced.}

\item{report.style}{\code{'concise'} style will produce the confidence
intervals for the f2 estimators; \code{'intermediate'} style will add
a list of individual f2s for all bootstrap samples; \code{detailed}
style will further add individual bootstrap samples along with their f2.
See Input/Output in Details.}

\item{f2.type}{Character strings indicating which f2 estimators should be
calculated.}

\item{ci.type}{Character strings indicating which type of  confidence
intervals should be estimated.}

\item{quantile.type}{Character strings indicating the type of percentile.}

\item{jackknife.type}{Character strings indicating the type for jackknife
method.}

\item{time.unit}{Character strings indicating the unit of time.
It should be either \code{"min"} for minute or \code{"h"} for hour.
It is mainly used for checking CV rules and making plot.
See Regulation in Details.}

\item{output.to.screen}{Logical. If \code{TRUE}, a concise style summary
report will be printed to screen.}

\item{sim.data.out}{Logical. If \code{TRUE}, all individual bootstrap data
sets will be included in the output.}
}
\value{
A list of 3 or 5 components.
\itemize{
\item \code{boot.ci}: A data frame of bootstrap f2 confidence intervals.
\item \code{boot.f2}: A data frame of all individual f2 values.
\item \code{boot.info}: A data frame with detailed information of bootstrap,
such as the time points used for calculation of f2 and the number of
\code{NA}s.
\item \code{boot.t} and \code{boot.r}: Individual bootstrap samples if
\code{sim.data.out = TRUE}.
}
}
\description{
Main function to estimate 90\% confidence intervals of f2 using bootstrap
methodology.
}
\examples{
\dontrun{
# time points
tp <- c(5, 10, 15, 20, 30, 45, 60)
# model.par for reference with low variability
par.r <- list(fmax = 100, fmax.cv = 3, mdt = 15, mdt.cv = 14,
              tlag = 0, tlag.cv = 0, beta = 1.5, beta.cv = 8)
# simulate reference data
dr <- sim.dp(tp, model.par = par.r, seed = 100, plot = FALSE)
# model.par for test
par.t <- list(fmax = 100, fmax.cv = 3, mdt = 12.29, mdt.cv = 12,
              tlag = 0, tlag.cv = 0, beta = 1.727, beta.cv = 9)
# simulate test data with low variability
dt <- sim.dp(tp, model.par = par.t, seed = 100, plot = FALSE)

# bootstrap
bootf2(dt$sim.disso, dr$sim.disso, n.boots = 100)
}

}
