% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcf2.R
\name{calcf2}
\alias{calcf2}
\title{Calculate Similarity Factor f2}
\usage{
calcf2(test, ref, regulation = c("EMA", "FDA", "WHO"),
       path.in, file.in, path.out, file.out, digits = 2L,
       cv.rule = TRUE, min.points = 3L, both.TR.85 = FALSE,
       f2.type = c("est.f2", "exp.f2", "bc.f2", "vc.exp.f2",
                   "vc.bc.f2", "all"), plot = TRUE,
       message = TRUE, time.unit = c("min", "h"),
       plot.start.time = 0, plot.max.unit = 24L)
}
\arguments{
\item{test, ref}{Data frames of dissolution profiles of test and reference
product if \code{path.in} and \code{file.in} are not specified; otherwise,
they should be character strings indicating the worksheet names of the
Excel file where the dissolution data is saved. Required format:
the first column should be time and the rest columns are dissolution
data of each unit. See Input/Output in Details.}

\item{regulation}{Character strings indicating regulatory guidelines.
See Regulation in Details.}

\item{path.in, file.in, path.out, file.out}{Character strings of input and
output directories and file names. See Input/Output in Details.}

\item{digits}{An integer indicating the decimal points for the output.}

\item{cv.rule}{Logical. If \code{TRUE}, CV rule will be checked
according to regulatory guidelines. See Regulation in Details.}

\item{min.points}{An integer indicating the minimum time points to be used
to calculate f2. Do not change the default value (3) for conventional f2
calculation. This parameter is mainly used for bootstrap f2 method.
See Regulation in Details.}

\item{both.TR.85}{Logical. If \code{TRUE}, the old (incorrect)
interpretation (that both the test and reference should release more than
85\%) will be used for f2 calculation when \code{regulation = 'FDA'}.
See Regulation in Details.}

\item{f2.type}{Character strings indicating which f2 estimators should be
calculated. For conventional f2 calculation, the default 'est.f2'
should be used. Other estimators are mainly for the bootstrap method.}

\item{plot}{Logical. If \code{TRUE}, a dissolution versus time plot will
be printed.}

\item{message}{Logical. If \code{TRUE}, the results and messages will be
printed on screen.}

\item{time.unit}{Character strings indicating the unit of time.
It should be either \code{"min"} for minute or \code{"h"} for hour.
It is mainly used for checking CV rules and making plot.
See Regulation in Details.}

\item{plot.start.time}{Numeric value indicating the start time for the plot.}

\item{plot.max.unit}{Integer. If the number of individual units is no more
than this value, mean and all individual profiles will be plotted,
otherwise, individual profiles will represented by boxplots.
Therefore, the value should not be too big.}
}
\value{
A vector of f2 and number of time points used for the calculation.
}
\description{
Main function to calculate f2 according to different regulatory guidelines.
}
\details{
\subsection{Input/Output}{
\itemize{
\item \code{path.in} is the directory where the input file is, and
\code{file.in} is the input data file. If they are both missing,
then the arguments \code{test} and \code{ref} must be specified,
and they should be R data frames, with 'time' in the first column and
dissolution data of individual units in the rest of the columns.
If only mean dissolution data is used, it should be the second column.
\item If \code{path.in} and \code{file.in} are specified, then \code{file.in}
should be the name of the Excel file where the dissolution data are saved.
The file name should include the extension 'xlsx' or 'xls', depending on
the version of the Excel. In that case, the argument \code{test} and
\code{ref} should be the names of the work sheets in the Excel,
given within the quotation marks.
\item \code{path.out} and \code{file.out} are the names of the output directory
and file. It is an overkill to output such simple calculations;
therefore, unless these two arguments are specified by the user,
results are printed on screen by default.
}
}

\subsection{Regulation}{

To use f2 method, different regulatory guidelines have slightly different
requirements. Some requirements are almost universal, such as
same time points for the test and reference product, minimum 3 time points
(excluding time zero), and twelve individual values for every time point for
for each formulation. Other requirements are different among different
regulatory guidelines, or at least interpreted differently. Two main issues
are the rules for the variability (CV Rule) and time points where
dissolution is more than 85\% (85\% Rule).
\subsection{CV rule}{
\itemize{
\item EMA: The CV of the first time point should not be greater than 20\%, and
the CV of the rest time points should not be greater than 10\%. The phrase
'the first time point' was later interpreted as all time points up to
10 min. For example, if there are 5 min and 10 min time points in the
dissolution profiles, the CV for 5 min and 10 min should not be great
than 20\%.
\item WHO: Same as EMA rule, i.e., CV not greater than 20\% for time points up to
10 min, and not greater than 10\% for the rest time points.
\item US FDA: More flexible. The CV for the early time points should not be
greater than 20\%, and for the rest time points, not greater than 10\%.
The phrase 'early time points' is typically interpreted as those points
up to 15 min, some times even up to 20 min according to unofficial
communication with FDA staff.
}
}

\subsection{85\% Rule}{
\itemize{
\item EMA: Dissolution profiles should be 'cut' at the time point where either
the test or reference release more than 85\%.
\item WHO: Dissolution profiles should be 'cut' at the time point where the
reference release more than 85\%.
\item US FDA: Only one measurement should be considered after 85\% dissolution
of both the products.
Many interpreted that the dissolution profiles should be 'cut'
at the time point where both the test and reference release more than 85\%.
However, this is a misunderstanding. Otherwise, when the slower profile
releases more than 85\%, there might be many measurements that are more than
85\% in the faster profile, which directly contradict to the statement that
only one measurement should be used after 85\% dissolution.
}
}

}
}
\examples{
tp <- c(5, 10, 15, 20, 30, 45, 60)

mod.par.t <- list(fmax = 100, fmax.cv = 2, tlag = 0, tlag.cv = 0,
                  mdt = 20, mdt.cv = 5, beta = 2.2, beta.cv = 5)

d.t <- sim.dp(tp, model.par = mod.par.t, seed = 100, n.units = 120L,
              plot = FALSE)

mod.par.r <- list(fmax = 100, fmax.cv = 2, tlag = 0, tlag.cv = 0,
                  mdt = 25, mdt.cv = 4, beta = 2.1, beta.cv = 3)

d.r <- sim.dp(tp, model.par = mod.par.r, seed = 100, n.units = 120L,
              plot = FALSE)

calcf2(d.t$sim.disso, d.r$sim.disso, plot = FALSE)

}
